// lib/types.ts
export type PageData = {
  title: string
  description: string
  image: string
  keywords: string
}

export type PageMeta = {
  home: PageData
  projects: PageData
  about: PageData
  contact: PageData
}